introduce login page and handle PMGQUAR tickets better
this moves most of the login logic to the login page, which will be
shown whenever the user is not logged in. Additionally the ticket refesh
loop will be stopped in case of a PMGQUAR ticket, as those do only
expire when they're older than the quarantine lifetime, so it's unlikely
that users will trigger that during browsing.

The refresh loop is problematic in that case since it will check the
ticket for validity which is hardcoded to be two hours after ticket
creation, which is much too low for quarantine tickets.

Instead of hardcoding a longer ticket lifetime for PMGQUAR tickets,
simply stop the refresh loop.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>