execute actions when opening gui from a link from quarantine mails
The PMG quarantine mails can contain links that should do some actions
directly, e.g. deliver or delete a mail. This works for our desktop gui
and worked in the old Framework7 mobile ui, but was forgotten to be
implemented in the new rust based gui.

This patch implements it in `SpamList` in the 'create' method.

To make it work properly we have to omit the `location.replace` when
doing a login, otherwise we lose the get parameters and refresh the
page. Instead we do a http_set_auth (like in create).

A SnackBar is also added on successful execution of an action.

Parts of this patch are take from Stoiko's first approach[0].

Note that while this patch only works with the catalog loader fix[1].
Without that, a ticket via login will be set, but since we don't
refresh the page anymore the catalog loader discards the state of the
login page and the 'on_login' callback is not called anymore which
means that the login page will still be shown.

0: https://lore.proxmox.com/pmg-devel/20251028163628.79739-1-s.ivanov@proxmox.com/
1: https://lore.proxmox.com/yew-devel/20251029143617.18134-1-d.csapak@proxmox.com/T/#u

Co-developed-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Reviewed-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Tested-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/20251030103209.1949668-1-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>